Move get_atlas_[n|c]ifti into utils.atlas.
